Сохраняет данные пользователя при включении/отключении Offers
