+++ package.json
@@ -10,8 +10,10 @@
     "start": "webpack-dev-server --open"
   },
   "devDependencies": {
+    "css-loader": "^5.0.1",
     "eslint": "7.13.0",
     "eslint-config-htmlacademy": "0.6.0",
+    "style-loader": "^2.0.0",
     "webpack": "4.44.2",
     "webpack-cli": "3.3.12",
     "webpack-dev-server": "3.11.0"
