+++ src/utils/const.js
@@ -21,52 +21,6 @@ export const FilterType = {
   PAST: `Past`,
 };
 
-export const typeWaypointOptionsMap = {
-  'taxi': [1, 2],
-  'bus': [2, 3, 4],
-  'train': [3, 5],
-  'ship': [2, 4, 5],
-  'transport': [1, 2, 3, 5],
-  'check-in': [1, 5],
-  'sightseeing': [1, 3, 5],
-  'restaurant': [2, 4, 5],
-  'flight': [1, 3],
-  'drive': [3, 4]
-};
-
-export const typeWaypointOptions = [{
-  value: `luggage`,
-  title: `Add luggage`,
-  price: 30,
-  checked: true,
-  kind: 1,
-}, {
-  value: `comfort`,
-  title: `Switch to comfort class`,
-  price: 100,
-  checked: true,
-  kind: 2
-}, {
-  value: `meal`,
-  title: `Add meal`,
-  price: 15,
-  checked: true,
-  kind: 3
-}, {
-  value: `seats`,
-  title: `Choose seats`,
-  price: 5,
-  checked: true,
-  kind: 4
-}, {
-  value: `train`,
-  title: `Travel by train`,
-  price: 40,
-  checked: true,
-  kind: 5
-}
-];
-
 export const destinationsMap = new Map([
   [`Amsterdam`, descriptions[1]],
   [`Chamonix`, descriptions[4] + descriptions[7]],
